qt: Replace icons for dark mode with 16x16 icons
* qt/icons/hint_dark.svg, qt/icons/password-generate_dark.svg,
qt/icons/visibility_dark.svg: Replace with copies of corresponding
normal icons with color changed to #eff0f1.
--

The old icons for dark mode were 22x22 sized icons which looked fuzzy
when used for the buttons. Using 16x16 sized icons ensures that the
icons for dark mode look as good as the icons for normal/light mode.
For qt5 the 16x16 sized icons are already used.

GnuPG-bug-id: 7230